requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
rest-framework-roles>=1.0.3<1.1
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.26.5,<0.27